refactor: extract receive concern into receive/ and cli/receive.py
Move ReceiveOptions and the zfs-receive command builder out of
snapshot/ into a dedicated receive/ package mirroring compress/
(type.py + command.py + __init__.py), and move the --receive- option
group decorator into cli/receive.py. snapshot/send.py now calls
receive.command(); ReceiveOptions is imported from ..receive.

This gives the receive value object a correct home (it was squatting in
snapshot/type.py) and sets up a symmetric send/ + cli/send.py split for
the coming send-flags work.

Refs #393.
